--- a/Makefile
+++ b/Makefile
@@ -1,3 +1,6 @@
+bench:
+	go test -bench .
+
 test:
 	go test -v ./...