--- a/asciitree.go
+++ b/asciitree.go
@@ -32,20 +32,20 @@ func NewFile(name string) *Node {
 //
 // Unlike AddFile and AddDir, Add returns the original node for
 // chaining.
-func (t *Node) Add(nodes ...*Node) *Node {
+func (n *Node) Add(nodes ...*Node) *Node {
 	for _, tree := range nodes {
-		t.Children = append(t.Children, tree)
+		n.Children = append(n.Children, tree)
 	}
-	return t
+	return n
 }
 
 // AddDir creates a tree node, forces it to be recognized as a directory,
 // and appends it to the node's children.
 //
 // Unlike AddDirs, AddDir returns the newly created node.
-func (t *Node) AddDir(name string) *Node {
+func (n *Node) AddDir(name string) *Node {
 	child := NewDir(name)
-	t.Children = append(t.Children, child)
+	n.Children = append(n.Children, child)
 	return child
 }
 
@@ -53,20 +53,20 @@ func (t *Node) AddDir(name string) *Node {
 // to be recognized as directories, and appends them to the node's children.
 //
 // Unlike AddDir, AddDirs returns the original node for chaining.
-func (t *Node) AddDirs(names ...string) *Node {
+func (n *Node) AddDirs(names ...string) *Node {
 	for _, name := range names {
 		child := NewDir(name)
-		t.Children = append(t.Children, child)
+		n.Children = append(n.Children, child)
 	}
-	return t
+	return n
 }
 
 // AddFile creates a tree node and appends it to the node's children.
 //
 // Unlike AddFiles, AddFile returns the newly created node.
-func (t *Node) AddFile(name string) *Node {
+func (n *Node) AddFile(name string) *Node {
 	child := NewFile(name)
-	t.Children = append(t.Children, child)
+	n.Children = append(n.Children, child)
 	return child
 }
 
@@ -74,46 +74,46 @@ func (t *Node) AddFile(name string) *Node {
 // them to the node's children.
 //
 // Unlike AddFile, AddFiles returns the original node for chaining.
-func (t *Node) AddFiles(names ...string) *Node {
+func (n *Node) AddFiles(names ...string) *Node {
 	for _, name := range names {
 		child := NewFile(name)
-		t.Children = append(t.Children, child)
+		n.Children = append(n.Children, child)
 	}
-	return t
+	return n
 }
 
 // Sort recursively sorts the node's children in place.
 //
 // Sort returns the original node for chaining.
-func (t *Node) Sort(opts ...SortOption) *Node {
+func (n *Node) Sort(opts ...SortOption) *Node {
 	options := newSortOptions(opts...)
-	sort.SliceStable(t.Children, func(i, j int) bool {
-		a := t.Children[i]
-		b := t.Children[j]
+	sort.SliceStable(n.Children, func(i, j int) bool {
+		a := n.Children[i]
+		b := n.Children[j]
 		if options.dirsFirst && a.IsDir && !b.IsDir {
 			return true
 		}
 		return a.Name < b.Name
 	})
-	for _, child := range t.Children {
+	for _, child := range n.Children {
 		child.Sort(opts...)
 	}
-	return t
+	return n
 }
 
 // String returns the tree's visual representation.
-func (t *Node) String() string {
+func (n *Node) String() string {
 	var builder strings.Builder
-	builder.WriteString(t.Name)
-	t.writeChildren(&builder, "")
+	builder.WriteString(n.Name)
+	n.writeChildren(&builder, "")
 	return builder.String()
 }
 
-func (t *Node) writeChildren(builder *strings.Builder, prefix string) {
-	for i, child := range t.Children {
+func (n *Node) writeChildren(builder *strings.Builder, prefix string) {
+	for i, child := range n.Children {
 		connector := "├── "
 		spacer := "│   "
-		if i == len(t.Children)-1 {
+		if i == len(n.Children)-1 {
 			connector = "└── "
 			spacer = "    "
 		}