--- a/bench_test.go
+++ b/bench_test.go
@@ -61,6 +61,28 @@ func BenchmarkAddFiles(b *testing.B) {
 	}
 }
 
+func BenchmarkNodeSort(b *testing.B) {
+	tree := NewDir("alfa").Add(
+		NewFile("bravo.txt"),
+		NewDir("charlie").Add(
+			NewFile("delta.txt"),
+			NewDir("echo").AddFiles("foxtrot.txt"),
+			NewDir("golf").Add(
+				NewDir("hotel\n[dir]\n[3 MB]").AddFiles(
+					"india.txt\n[file]\n[1 MB]",
+					"juliet.txt\n[file]\n[2 MB]",
+				),
+			),
+		),
+		NewDir("kilo").AddFiles("lima.txt"),
+		NewDir("mike"),
+		NewFile("november.txt"),
+	)
+	for i := 0; i < b.N; i++ {
+		tree.Sort()
+	}
+}
+
 func BenchmarkNodeString(b *testing.B) {
 	tree := NewDir("alfa").Add(
 		NewFile("bravo.txt"),