This PR benchmarks some of the heavier functions and optimizes them.

Before:

```
cpu: Intel(R) Core(TM) i7-4980HQ CPU @ 2.80GHz
BenchmarkNodeAdd-8      	 3905551	       300.2 ns/op
BenchmarkAddDirs-8      	 1657606	       719.6 ns/op
BenchmarkAddFiles-8     	 1607150	       742.3 ns/op
BenchmarkNodeSort-8     	  586012	      2095 ns/op
BenchmarkNodeString-8   	  511766	      2334 ns/op
```

After:

```
cpu: Intel(R) Core(TM) i7-4980HQ CPU @ 2.80GHz
BenchmarkNodeAdd-8      	14195162	        86.82 ns/op
BenchmarkAddDirs-8      	 2131742	       565.4 ns/op
BenchmarkAddFiles-8     	 2171773	       556.3 ns/op
BenchmarkNodeSort-8     	  766713	      1552 ns/op
BenchmarkNodeString-8   	  906979	      1336 ns/op
```